+++ src/dmd_era5/era5_download/era5_download.py
@@ -7,7 +7,7 @@ import xarray as xr
 from dvc.repo import Repo as DvcRepo
 from pyprojroot import here
 
-from dmd_era5.core import config_parser, config_reader, log_and_print, setup_logger
+from dmd_era5 import config_parser, config_reader, log_and_print, setup_logger
 from dmd_era5.create_mock_data import create_mock_era5
 from dmd_era5.dvc_tools import add_data_to_dvc, retrieve_data_from_dvc
 from dmd_era5.slice_tools import resample_era5_dataset, slice_era5_dataset
